[rx] make observable transaction rather a single result. fix issue where insert called twice on wrapper object. fix issue where db not deleted on destroy. fix broken unit tests.
