+++ dbflow-tests/src/test/java/com/raizlabs/dbflow5/provider/ContentProviderTests.kt
@@ -32,96 +32,102 @@ class ContentProviderTests : BaseUnitTest() {
     }
 
     @Test
-    fun testContentProviderUtils() = database(ContentDatabase::class) {
-        tables(NoteModel::class.java, ContentProviderModel::class.java)
+    fun testContentProviderUtils() {
+        database(ContentDatabase::class) {
+            tables(NoteModel::class.java, ContentProviderModel::class.java)
 
-        var contentProviderModel = ContentProviderModel()
-        contentProviderModel.notes = "Test"
-        var uri = ContentUtils.insert(mockContentResolver, TestContentProvider.ContentProviderModel.CONTENT_URI, contentProviderModel)
-        assertEquals(TestContentProvider.ContentProviderModel.CONTENT_URI.toString() + "/" + contentProviderModel.id, uri.toString())
-        assertTrue(contentProviderModel.exists())
+            var contentProviderModel = ContentProviderModel()
+            contentProviderModel.notes = "Test"
+            var uri = ContentUtils.insert(mockContentResolver, TestContentProvider.ContentProviderModel.CONTENT_URI, contentProviderModel)
+            assertEquals(TestContentProvider.ContentProviderModel.CONTENT_URI.toString() + "/" + contentProviderModel.id, uri.toString())
+            assertTrue(contentProviderModel.exists())
 
-        contentProviderModel.notes = "NewTest"
-        val update = ContentUtils.update(mockContentResolver, TestContentProvider.ContentProviderModel.CONTENT_URI, contentProviderModel)
-        assertEquals(update.toLong(), 1)
-        assertTrue(contentProviderModel.exists())
-        contentProviderModel = contentProviderModel.load(this)!!
-        assertEquals("NewTest", contentProviderModel.notes)
+            contentProviderModel.notes = "NewTest"
+            val update = ContentUtils.update(mockContentResolver, TestContentProvider.ContentProviderModel.CONTENT_URI, contentProviderModel)
+            assertEquals(update.toLong(), 1)
+            assertTrue(contentProviderModel.exists())
+            contentProviderModel = contentProviderModel.load(this)!!
+            assertEquals("NewTest", contentProviderModel.notes)
 
-        val noteModel = NoteModel()
-        noteModel.note = "Test"
-        noteModel.contentProviderModel = contentProviderModel
-        uri = ContentUtils.insert(mockContentResolver, TestContentProvider.NoteModel.CONTENT_URI, noteModel)
-        assertEquals(TestContentProvider.NoteModel.CONTENT_URI.toString() + "/" + noteModel.id, uri.toString())
-        assertTrue(noteModel.exists())
+            val noteModel = NoteModel()
+            noteModel.note = "Test"
+            noteModel.contentProviderModel = contentProviderModel
+            uri = ContentUtils.insert(mockContentResolver, TestContentProvider.NoteModel.CONTENT_URI, noteModel)
+            assertEquals(TestContentProvider.NoteModel.CONTENT_URI.toString() + "/" + noteModel.id, uri.toString())
+            assertTrue(noteModel.exists())
 
-        assertTrue(ContentUtils.delete(mockContentResolver, TestContentProvider.NoteModel.CONTENT_URI, noteModel) > 0)
-        assertTrue(!noteModel.exists())
+            assertTrue(ContentUtils.delete(mockContentResolver, TestContentProvider.NoteModel.CONTENT_URI, noteModel) > 0)
+            assertTrue(!noteModel.exists())
 
-        assertTrue(ContentUtils.delete(mockContentResolver, TestContentProvider.ContentProviderModel.CONTENT_URI, contentProviderModel) > 0)
-        assertTrue(!contentProviderModel.exists())
+            assertTrue(ContentUtils.delete(mockContentResolver, TestContentProvider.ContentProviderModel.CONTENT_URI, contentProviderModel) > 0)
+            assertTrue(!contentProviderModel.exists())
 
-        tables(NoteModel::class.java, ContentProviderModel::class.java)
+            tables(NoteModel::class.java, ContentProviderModel::class.java)
+        }
     }
 
     @Test
-    fun testContentProviderNative() = database(ContentDatabase::class) {
-        tables(NoteModel::class.java, ContentProviderModel::class.java)
+    fun testContentProviderNative() {
+        database(ContentDatabase::class) {
+            tables(NoteModel::class.java, ContentProviderModel::class.java)
 
-        var contentProviderModel = ContentProviderModel(notes = "Test")
-        contentProviderModel.insert()
-        assertTrue(contentProviderModel.exists())
+            var contentProviderModel = ContentProviderModel(notes = "Test")
+            contentProviderModel.insert()
+            assertTrue(contentProviderModel.exists())
 
-        contentProviderModel.notes = "NewTest"
-        contentProviderModel.update()
-        contentProviderModel = contentProviderModel.load(this)!!
-        assertEquals("NewTest", contentProviderModel.notes)
+            contentProviderModel.notes = "NewTest"
+            contentProviderModel.update()
+            contentProviderModel = contentProviderModel.load(this)!!
+            assertEquals("NewTest", contentProviderModel.notes)
 
-        var noteModel = NoteModel(note = "Test", contentProviderModel = contentProviderModel)
-        noteModel.insert()
+            var noteModel = NoteModel(note = "Test", contentProviderModel = contentProviderModel)
+            noteModel.insert()
 
-        noteModel.note = "NewTest"
-        noteModel.update()
-        noteModel = noteModel.load(this)!!
-        assertEquals("NewTest", noteModel.note)
+            noteModel.note = "NewTest"
+            noteModel.update()
+            noteModel = noteModel.load(this)!!
+            assertEquals("NewTest", noteModel.note)
 
-        assertTrue(noteModel.exists())
+            assertTrue(noteModel.exists())
 
-        noteModel.delete()
-        assertTrue(!noteModel.exists())
+            noteModel.delete()
+            assertTrue(!noteModel.exists())
 
-        contentProviderModel.delete()
-        assertTrue(!contentProviderModel.exists())
+            contentProviderModel.delete()
+            assertTrue(!contentProviderModel.exists())
 
-        tables(NoteModel::class.java, ContentProviderModel::class.java)
+            tables(NoteModel::class.java, ContentProviderModel::class.java)
+        }
     }
 
     @Test
-    fun testSyncableModel() = database(ContentDatabase::class) {
-        table(this, TestSyncableModel::class.java)
+    fun testSyncableModel() {
+        database(ContentDatabase::class) {
+            table(this, TestSyncableModel::class.java)
 
-        var testSyncableModel = TestSyncableModel(name = "Name")
-        testSyncableModel.save()
+            var testSyncableModel = TestSyncableModel(name = "Name")
+            testSyncableModel.save()
 
-        assertTrue(testSyncableModel.exists())
+            assertTrue(testSyncableModel.exists())
 
-        testSyncableModel.name = "TestName"
-        testSyncableModel.update()
-        assertEquals(testSyncableModel.name, "TestName")
+            testSyncableModel.name = "TestName"
+            testSyncableModel.update()
+            assertEquals(testSyncableModel.name, "TestName")
 
-        testSyncableModel = (select from TestSyncableModel::class
-            where (TestSyncableModel_Table.id.`is`(testSyncableModel.id))).result!!
+            testSyncableModel = (select from TestSyncableModel::class
+                where (TestSyncableModel_Table.id.`is`(testSyncableModel.id))).result!!
 
-        var fromContentProvider = TestSyncableModel(id = testSyncableModel.id)
-        fromContentProvider = fromContentProvider.load(this)!!
+            var fromContentProvider = TestSyncableModel(id = testSyncableModel.id)
+            fromContentProvider = fromContentProvider.load(this)!!
 
-        assertEquals(fromContentProvider.name, testSyncableModel.name)
-        assertEquals(fromContentProvider.id, testSyncableModel.id)
+            assertEquals(fromContentProvider.name, testSyncableModel.name)
+            assertEquals(fromContentProvider.id, testSyncableModel.id)
 
-        testSyncableModel.delete()
-        assertFalse(testSyncableModel.exists())
+            testSyncableModel.delete()
+            assertFalse(testSyncableModel.exists())
 
-        table(this, TestSyncableModel::class.java)
+            table(this, TestSyncableModel::class.java)
+        }
     }
 
 }
